Added remote client and synchronous server operation
A remote controller can now execute wt1 (Worker Task) and various at (Assembly Tasks) by sending http requests to the server.
The controller is implemented in Bash and cURL currently. For the requests to be executed sequentially, the execSync function
of 'child-process' NodeJs module is used, this makes the robot always available when a request comes to the server.

Todo: Make the server for the workbenches. Implement composite uS placePart correctly. Have the robotic arm screw the leg.

Note: Some functionalities are moved to the workbench gripper and will be implemented in their server. E.g. when a leg is placed
into the leg base, the workbench gripper (which holds the leg base) will turn and the robotic arm will repeat the process with
the same coordinates for all the legs.