Reworked place part uS to work as a composite of primitive ones and fixed server
The place part uS is now a bash file that the server runs. It now consists of some primitive uS.
The server was changed to implement this command.
Server bug fix:
execSync in nodeJS didn't seem to run the GetPose uS and had to be changed to asynchronus exec.
This could've been because this service doesn't use the dobot queue, but seems like the Alarm reset
is working fine. It remains a mystery, but it works now.